requirements.txt
selenium
pyyaml